requirements.txt
selenium
pyyaml